requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
numpy>=1.24.3
plotly>=5.15.0
networkx>=3.1
torch>=2.0.0
transformers>=4.30.2
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
aioboto3
oracledb
nano-graphrag
lightrag-hku
chromadb
PyPDF2
docling